fixing project  floder structure
