+++ include/interaction/HelperFunctions.h
