+++ libs/transferfunction/embedded_colormaps.h
