+++ libs/transferfunction/transfer_function_multichannel_widget.h
