+++ libs/transferfunction/transfer_function_widget.h
