+++ src/render/FontHandler.cpp
@@ -3,7 +3,7 @@
 #endif
 #include <GL/gl.h>
 
-#include "FontHandler.h"
+#include "../../include/render/FontHandler.h"
 #include <iostream>
 #include <time.h>
 #include <iomanip>
