+++ src/render/Volume.cpp
@@ -30,7 +30,7 @@
 #define _CRT_SECURE_NO_WARNINGS
 #endif
 
-#include "Volume.h"
+#include "../../include/render/Volume.h"
 #include <iostream>
 
 Volume::Volume(unsigned int width, unsigned int height, unsigned int depth, double x_scale, double y_scale, double z_scale, unsigned int datatypesize, unsigned int channel)
